+++ R/read_data.R
@@ -180,6 +180,21 @@ get_chunk_size <- function(datatype, dimensions) {
   return(as.integer(buffer_size))
 }
 
+#' Determine the size of chunk in bytes after decompression
+#' 
+#' @param datatype A list of details for the array datatype.  Expected to be
+#' produced by [.parse_datatype()].
+#' @param dimensions A list containing the dimensions of the chunk.  Expected 
+#' to be found in a list produced by [read_array_metadata()].
+#' 
+#' @returns An integer giving the size of the chunk in bytes
+#'   
+#' @keywords Internal
+get_decompressed_chunk_size <- function(datatype, dimensions) {
+    buffer_size <- prod(unlist(dimensions), datatype$nbytes)
+    return(as.integer(buffer_size))
+}
+
 #' Read a single Zarr chunk
 #'
 #' @param zarr_array_path A character vector of length 1, giving the path to the
@@ -355,7 +370,7 @@ read_chunk <- function(zarr_array_path, chunk_id, metadata, s3_client = NULL,
 .decompress_chunk <- function(compressed_chunk, metadata) {
   decompressor <- metadata$compressor$id
   datatype <- metadata$datatype
-  buffer_size <- get_chunk_size(datatype, dimensions = metadata$chunks)
+  buffer_size <- get_decompressed_chunk_size(datatype, dimensions = metadata$chunks)
 
   if(is.null(decompressor)) {
     decompressed_chunk <- compressed_chunk